tweaks: change an integer to a short, and reshuffle it for better packing
A short allows for more than 32 thousand values, and the maximum number
of multiline regexes in any of the current syntaxes is... just four.

Reshuffle it to the beginning also because it is used the most often.

+++ src/color.c
@@ -266,7 +266,7 @@ void set_up_multicache(linestruct *line)
 {
 	line->multidata = (short *)nmalloc(openfile->syntax->nmultis * sizeof(short));
 
-	for (int index = 0; index < openfile->syntax->nmultis; index++)
+	for (short index = 0; index < openfile->syntax->nmultis; index++)
 		line->multidata[index] = -1;
 }
